Reorganized the folder structure
