+++ go1percent/helpers/Go1PercentFEAutomation/TECHHUB/AdminUser/go1PercentTechhub.js
@@ -1,7 +1,5 @@
 module.exports = {
 
-
-
     "Navigate to All Techhubs page": function (browser) {
         browser
             .page.techhub()
